% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{categorizeNonNumericColumns}
\alias{categorizeNonNumericColumns}
\title{Categorize Non-numeric Columns}
\usage{
categorizeNonNumericColumns(df, columns = NULL, suffix = "_ctgr")
}
\arguments{
\item{df}{a data frame}

\item{columns}{a vector with the column names to be categorized.}

\item{suffix}{the suffix to be appeded to the new column names}
}
\value{
a data frame with the original and the discretized
}
\description{
Categorizes the non-numeric columns of the given data frame \code{df}.
The non-numeric columns will be converted to factor columns with numeric labels.
This is useful to machine learning algorithms that use numbers instead of texts.
}
\examples{
categorizeNonNumericColumns(startupssmall)
categorizeNonNumericColumns(startupssmall, c("State"))
categorizeNonNumericColumns(startupssmall, c("State"), suffix="_categorized")
}
