% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{dataQuantiles}
\alias{dataQuantiles}
\title{Data Quantiles}
\usage{
dataQuantiles(data, n = 4, na.rm = TRUE)
}
\arguments{
\item{data}{a data vector}

\item{n}{the number of groups to divide the data vector}

\item{na.rm}{logical; if true, any NA
and \code{NaN}'s are removed from \code{data} before the quantiles are computed.}
}
\value{
a named numeric vector value for 0%, 1/n%,...,n/n%
}
\description{
Retrieves the 0 up to 100 quantiles that split the vector \code{data} in \code{n} groups.
}
\examples{
dataQuantiles(rep(1:10,10), 1)
dataQuantiles(rep(1:10,10), 3)
dataQuantiles(rep(seq(from=0,to=10,by=0.1),10), 4)
}
