% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{discretizeByQuantiles}
\alias{discretizeByQuantiles}
\title{Discretize Data by Quantiles}
\usage{
discretizeByQuantiles(data, n = 4, na.rm = TRUE)
}
\arguments{
\item{data}{a data vector}

\item{n}{the number of groups to divide the data vector}

\item{na.rm}{logical; if true, any NA
and \code{NaN}'s are removed from \code{data} before the quantiles are computed.}
}
\value{
a vector where the numerical data are replaced by its equivalent groups
}
\description{
Discretizes the given \code{data}, splitting it in \code{n} groups
based on its quantiles.
}
\examples{
discretizeByQuantiles(rep(1:10,10))
discretizeByQuantiles(rep(1:10,10), 1)
discretizeByQuantiles(rep(1:10,10), 3)
discretizeByQuantiles(rep(seq(from=0,to=10,by=0.1),10))
discretizeByQuantiles(rep(seq(from=0,to=10,by=0.1),10), 5)
}
