% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{discretizeColumnsByQuantiles}
\alias{discretizeColumnsByQuantiles}
\title{Discretize Numeric Columns by Quantile}
\usage{
discretizeColumnsByQuantiles(df, columns = NULL, ns = NULL,
  suffix = "_dscr")
}
\arguments{
\item{df}{a data frame}

\item{columns}{a vector with the column names to be discretized. If \code{is.null(columns)},
every numeric column of \code{df} will be discretized}

\item{ns}{a vector containing the number of groups to discretize each column specified by \code{columns}. If
the length of \code{ns} is different from the length of \code{columns} or \code{is.null(ns)}, then it will be 
used a vector of 4s with the same size of \code{columns}}

\item{suffix}{the suffix to be appeded to the new column names}
}
\value{
a data frame with the original and the discretized
}
\description{
Discretizes the given data frame \code{df}, splitting its columns \code{columns[0..i]} in \code{ns[0..i]} groups.
The new columns will have the same name of the original columns including the \code{suffix}, and will be appeded in the end of the \code{df}.
Processed the data \code{startupsbig} in ~1min and \code{startupshuge} in ~10min.
}
\examples{
discretizeColumnsByQuantiles(startupssmall)
discretizeColumnsByQuantiles(startupssmall, c("R.D.Spend","Profit"), c(3,5))
}
