% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_factor_conversion.R
\name{func_factor_conversion}
\alias{func_factor_conversion}
\title{Utility function to convert to factors
Conversation Function}
\usage{
func_factor_conversion(var_name, data_name, question_type)
}
\arguments{
\item{var_name}{Variable name of interest}

\item{data_name}{Dataset to be used}

\item{question_type}{Question type of the variable}
}
\value{

}
\description{
Utility function to convert to factors
Conversation Function
}
\examples{
#func_factor_conversion(var_name, data_name, question_type)
}
