% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_split_to_rows.R
\name{func_split_to_rows}
\alias{func_split_to_rows}
\title{Title}
\usage{
func_split_to_rows(dt_name, question_type, var_name, sub_form_column)
}
\arguments{
\item{dt_name}{data set}

\item{question_type}{question type e.g. Single Choice}

\item{var_name}{variable name in subform}

\item{sub_form_column}{sub form column you are picking from \verb{matrix codes} or \code{codes}}
}
\value{

}
\description{
Title
}
\examples{
# func_split_to_rows(dt_name = code_working,
}
