% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_subform.R
\name{func_subform}
\alias{func_subform}
\title{Utility Function to handle subform generation options}
\usage{
func_subform()
}
\value{

}
\description{
Utility Function to handle subform generation options
}
\examples{
#func_subform()
}
