% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_unite.R
\name{func_unite}
\alias{func_unite}
\title{Merges variables into one separated by comma, the variables should have NA}
\usage{
func_unite(dataset_name, var_name)
}
\arguments{
\item{dataset_name}{input data set}

\item{var_name}{name of the variable that you need to merge into one}
}
\value{

}
\description{
Merges variables into one separated by comma, the variables should have NA
}
\examples{
#fun_unite(dt, var_name = "var1")
}
