% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_var_lab.R
\name{func_var_lab}
\alias{func_var_lab}
\title{Utility function to create variable label}
\usage{
func_var_lab(code_working, question_type, var_name, loop_consideration = FALSE)
}
\arguments{
\item{code_working}{manipulated subform containing all the ordered questions}

\item{question_type}{Question type for particular question}

\item{var_name}{Variable name of interest}

\item{loop_consideration}{whether there is any loop or not}
}
\value{

}
\description{
Utility function to create variable label
}
\examples{
#func_var_lab(code_working, question_type, var_name, loop_consideration = FALSE)
}
