% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_var_lab_spss.R
\name{func_var_lab_spss}
\alias{func_var_lab_spss}
\title{Utility function to handle var labelling to spss}
\usage{
func_var_lab_spss(var_name, data_name, question_type)
}
\arguments{
\item{var_name}{Variable name in focus}

\item{data_name}{Data set in focus}

\item{question_type}{Question type to be used}
}
\value{

}
\description{
Part that does variable labelling
}
\examples{
#func_var_lab_spss(var_name, data_name, question_type)
}
