% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{getNonNumericColumns}
\alias{getNonNumericColumns}
\title{Non-numeric Columns}
\usage{
getNonNumericColumns(df)
}
\arguments{
\item{df}{the dataframe}
}
\value{
a vector
}
\description{
Retrieves a vector with the names of the non-numeric columns of a dataframe \code{df}
}
\examples{
getNonNumericColumns(startupssmall)
}
