% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{getNumericColumns}
\alias{getNumericColumns}
\title{Numeric Columns}
\usage{
getNumericColumns(df)
}
\arguments{
\item{df}{the dataframe}
}
\value{
a vector
}
\description{
Retrieves a vector with the names of the numeric columns of a dataframe \code{df}
}
\examples{
getNumericColumns(startupssmall)
}
