% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{histQuantiles}
\alias{histQuantiles}
\title{Data Histogram by Quantiles}
\usage{
histQuantiles(name = "data", data, n = 4, na.rm = TRUE, plot = TRUE)
}
\arguments{
\item{name}{a name for the data. default="data"}

\item{data}{a data vector}

\item{n}{the number of groups to divide the data vector}

\item{na.rm}{logical; if true, any NA
and \code{NaN}'s are removed from \code{data} before the quantiles are computed.}

\item{plot}{logical; if true, the histogram is plotted}
}
\value{
the histogram
}
\description{
Creates a histogram of \code{name} with the given \code{data}, splitting it in \code{n} groups
based on its quantiles.
}
\examples{
histQuantiles("Grades", rep(1:10,10), 1)
histQuantiles("Grades", rep(1:10,10), 3)
histQuantiles("Grades", rep(seq(from=0,to=10,by=0.1),10), 4)

}
