% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{preprocess}
\alias{preprocess}
\title{Data Preprocessing}
\usage{
preprocess(csv, csvSuffix = "_preprocessed", numericColumns = NULL,
  ns = 4, naBehavior = 2, numericColumnsSuffix = "_dscr",
  nonNumericColumns = NULL, nonNumericColumnsSuffix = "_ctgr",
  rmNonNumericColumns = TRUE)
}
\arguments{
\item{csv}{a csv file path.}

\item{csvSuffix}{The suffix that will be included after the name of the new csv.}

\item{numericColumns}{a vector with the numeric column names to be discretized}

\item{ns}{a vector containing the number of groups to discretize each column specified by \code{columns}}

\item{naBehavior}{integer; depicts what to do whenever a \code{NaN} is found in a numerical column. 0=remove line, 1=mean, 2=median, 3=mode.}

\item{numericColumnsSuffix}{the suffix to be appended to new numeric columns names}

\item{nonNumericColumns}{a vector with the non-numeric column names to be categorized}

\item{nonNumericColumnsSuffix}{the suffix to be appended to new non-numeric columns names}

\item{rmNonNumericColumns}{boolean; if \code{TRUE}}
}
\value{
the saved file
}
\description{
Preprocess the data of a csv file and generates a new file with the original
and the preprocessed data.
}
\examples{
\dontrun{
preprocess("data.csv")
preprocess(csv="data.csv", numericColumns=c("A", "B"), ns=c(3, 5))
preprocess(csv="data.csv", naBehavior=0, numericColumns=c("A", "B"), ns=c(3, 5))
}
}
