% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{preprocessNaN}
\alias{preprocessNaN}
\title{Process \code{NaN}s}
\usage{
preprocessNaN(df, naBehavior, columns = NULL)
}
\arguments{
\item{df}{a data frame}

\item{naBehavior}{a vector of integers; depicts what to do whenever a \code{NaN} is found in a numerical column. 0=remove line, 1=mean, 2=median, 3=mode.}

\item{columns}{a vector with the columns names}
}
\value{
a data frame with the processed \code{NaN}s
}
\description{
Process the \code{NaN}s of the data frame, by removing rows or replacing them by mean, median or mode.
}
\examples{
preprocessNaN(startupssmall, 0)
preprocessNaN(startupssmall, 2, c("R.D.Spend","Profit"))
}
