% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_func.R
\name{spss_func}
\alias{spss_func}
\title{SPSS Variable naming convention change}
\usage{
spss_func(var_naming, trunc = TRUE)
}
\arguments{
\item{var_naming}{variable name in focus}

\item{trunc}{whether we are truncating the characters to 64}
}
\value{

}
\description{
Removing special characters
}
\examples{
#spss_func(var_naming, trunc = TRUE)
}
