% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{startupsbig}
\alias{startupsbig}
\title{Data from some fictional startups. The data of startupssmall was replicated so as to
have a bigger data frame.}
\format{A data frame with 3,276,800 rows and 5 variables:
\describe{
  \item{R.D.Spend}{expenses in research and development, in US dollars}
  \item{Administration}{expenses in administration, in US dollars}
  \item{Marketing.Spend}{expenses in marketing, in US dollars}
  \item{State}{the state of the startup}
  \item{Profit}{the profit, int US dollars}
}}
\source{
\url{https://www.udemy.com/machinelearning/}
}
\usage{
data(startupsbig)
}
\description{
Data from some fictional startups. The data of startupssmall was replicated so as to
have a bigger data frame.
}
\keyword{data}
