% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{startupshuge}
\alias{startupshuge}
\title{Data from some fictional startups. The data of startupssmall was replicated so as to
have a bigger data frame.}
\format{A data frame with 3,276,800 rows and 50 variables:
\describe{
  \item{R.D.Spend}{expenses in research and development, in US dollars}
  \item{Administration}{expenses in administration, in US dollars}
  \item{Marketing.Spend}{expenses in marketing, in US dollars}
  \item{State}{the state of the startup}
  \item{Profit}{the profit, int US dollars}
  \item{R.D.Spend_1}{expenses in research and development, in US dollars}
  \item{Administration_1}{expenses in administration, in US dollars}
  \item{Marketing.Spend_1}{expenses in marketing, in US dollars}
  \item{State_1}{the state of the startup}
  \item{Profit_1}{the profit, int US dollars}
  \item{R.D.Spend_2}{expenses in research and development, in US dollars}
  \item{Administration_2}{expenses in administration, in US dollars}
  \item{Marketing.Spend_2}{expenses in marketing, in US dollars}
  \item{State_2}{the state of the startup}
  \item{Profit_2}{the profit, int US dollars}
  \item{R.D.Spend_3}{expenses in research and development, in US dollars}
  \item{Administration_3}{expenses in administration, in US dollars}
  \item{Marketing.Spend_3}{expenses in marketing, in US dollars}
  \item{State_3}{the state of the startup}
  \item{Profit_3}{the profit, int US dollars}
  \item{R.D.Spend_4}{expenses in research and development, in US dollars}
  \item{Administration_4}{expenses in administration, in US dollars}
  \item{Marketing.Spend_4}{expenses in marketing, in US dollars}
  \item{State_4}{the state of the startup}
  \item{Profit_4}{the profit, int US dollars}
  \item{R.D.Spend_5}{expenses in research and development, in US dollars}
  \item{Administration_5}{expenses in administration, in US dollars}
  \item{Marketing.Spend_5}{expenses in marketing, in US dollars}
  \item{State_5}{the state of the startup}
  \item{Profit_5}{the profit, int US dollars}
  \item{R.D.Spend_6}{expenses in research and development, in US dollars}
  \item{Administration_6}{expenses in administration, in US dollars}
  \item{Marketing.Spend_6}{expenses in marketing, in US dollars}
  \item{State_6}{the state of the startup}
  \item{Profit_6}{the profit, int US dollars}
  \item{R.D.Spend_7}{expenses in research and development, in US dollars}
  \item{Administration_7}{expenses in administration, in US dollars}
  \item{Marketing.Spend_7}{expenses in marketing, in US dollars}
  \item{State_7}{the state of the startup}
  \item{Profit_7}{the profit, int US dollars}
  \item{R.D.Spend_8}{expenses in research and development, in US dollars}
  \item{Administration_8}{expenses in administration, in US dollars}
  \item{Marketing.Spend_8}{expenses in marketing, in US dollars}
  \item{State_8}{the state of the startup}
  \item{Profit_8}{the profit, int US dollars}
  \item{R.D.Spend_9}{expenses in research and development, in US dollars}
  \item{Administration_9}{expenses in administration, in US dollars}
  \item{Marketing.Spend_9}{expenses in marketing, in US dollars}
  \item{State_9}{the state of the startup}
  \item{Profit_9}{the profit, int US dollars}
}}
\source{
\url{https://www.udemy.com/machinelearning/}
}
\usage{
data(startupshuge)
}
\description{
Data from some fictional startups. The data of startupssmall was replicated so as to
have a bigger data frame.
}
\keyword{data}
