\name{data}
\alias{data}
\docType{data}
\title{
Example data (particle number concentrations in SMEAR I Varrio forest)
}
\description{
The raw data is downloaded from https://avaa.tdata.fi/web/smart/smear/download.
}
\usage{data}
\format{
  A data frame with 7640 observations on the following 65 variables.
  \describe{
    \item{\code{date}}{a POSIXct}
    \item{\code{tconc}}{a numeric vector}
    \item{\code{TPNC}}{a numeric vector}
    \item{\code{monthyear}}{a character vector}
    \item{\samp{1}}{a numeric vector}
    \item{\samp{1.12}}{a numeric vector}
    \item{\samp{1.26}}{a numeric vector}
    \item{\samp{1.41}}{a numeric vector}
    \item{\samp{1.58}}{a numeric vector}
    \item{\samp{1.78}}{a numeric vector}
    \item{\samp{2}}{a numeric vector}
    \item{\samp{2.24}}{a numeric vector}
    \item{\samp{2.51}}{a numeric vector}
    \item{\samp{2.82}}{a numeric vector}
    \item{\samp{3.16}}{a numeric vector}
    \item{\samp{3.55}}{a numeric vector}
    \item{\samp{3.98}}{a numeric vector}
    \item{\samp{4.47}}{a numeric vector}
    \item{\samp{5.01}}{a numeric vector}
    \item{\samp{5.62}}{a numeric vector}
    \item{\samp{6.31}}{a numeric vector}
    \item{\samp{7.08}}{a numeric vector}
    \item{\samp{7.94}}{a numeric vector}
    \item{\samp{8.91}}{a numeric vector}
    \item{\samp{10}}{a numeric vector}
    \item{\samp{11.2}}{a numeric vector}
    \item{\samp{12.6}}{a numeric vector}
    \item{\samp{14.1}}{a numeric vector}
    \item{\samp{15.8}}{a numeric vector}
    \item{\samp{17.8}}{a numeric vector}
    \item{\samp{20}}{a numeric vector}
    \item{\samp{22.4}}{a numeric vector}
    \item{\samp{25.1}}{a numeric vector}
    \item{\samp{28.2}}{a numeric vector}
    \item{\samp{31.6}}{a numeric vector}
    \item{\samp{35.5}}{a numeric vector}
    \item{\samp{39.8}}{a numeric vector}
    \item{\samp{44.7}}{a numeric vector}
    \item{\samp{50.1}}{a numeric vector}
    \item{\samp{56.2}}{a numeric vector}
    \item{\samp{63.1}}{a numeric vector}
    \item{\samp{70.8}}{a numeric vector}
    \item{\samp{79.4}}{a numeric vector}
    \item{\samp{89.1}}{a numeric vector}
    \item{\samp{100}}{a numeric vector}
    \item{\samp{112}}{a numeric vector}
    \item{\samp{126}}{a numeric vector}
    \item{\samp{141}}{a numeric vector}
    \item{\samp{158}}{a numeric vector}
    \item{\samp{178}}{a numeric vector}
    \item{\samp{200}}{a numeric vector}
    \item{\samp{224}}{a numeric vector}
    \item{\samp{251}}{a numeric vector}
    \item{\samp{282}}{a numeric vector}
    \item{\samp{316}}{a numeric vector}
    \item{\samp{355}}{a numeric vector}
    \item{\samp{398}}{a numeric vector}
    \item{\samp{447}}{a numeric vector}
    \item{\samp{501}}{a numeric vector}
    \item{\samp{562}}{a numeric vector}
    \item{\samp{631}}{a numeric vector}
    \item{\samp{708}}{a numeric vector}
    \item{\samp{794}}{a numeric vector}
    \item{\samp{891}}{a numeric vector}
    \item{\samp{1000}}{a numeric vector}
  }
}

\source{
https://avaa.tdata.fi/web/smart/smear/download
}
\references{
1. Example data is from https://smear.avaa.csc.fi/download. It includes particle number concentrations in SMEAR I Varrio forest.
}
\examples{
data
## maybe str(data)
}
\keyword{datasets}
